Updated save function to categorys outputted from API, and corresponding unittest

+++ tests/test.py
@@ -28,19 +28,26 @@ class TestSavingRoutine(unittest.TestCase):
     def test_save_routine(self):
         test_file = "test_user_routines.csv"
         test_data = [
-            {"Category": "Cleanser", "Name": "CeraVe Hydrating Cleanser", "Price": "9.99"},
-            {"Category": "MakeUp", "Name": "MAC Ruby Woo Lipstick", "Price": "19.99"}
+            {"Brand": "Fenty", "Product": "Foundation", "Price": "30", "Description": "Test description"},
+            {"Brand": "MAC", "Product": "Lipstick", "Price": "19.99", "Description": "MAC Ruby Woo Lipstick"}
         ]
         save_routine(test_data, filename=test_file)
         with open(test_file, "r") as csv_file:
             reader = csv.DictReader(csv_file)
             rows = list(reader)
+            for row in rows:
+                print(row)
         self.assertEqual(len(rows), 2)
-        self.assertEqual(rows[0]["Name"], "CeraVe Hydrating Cleanser")
+        self.assertEqual(rows[0]["Product"], "Foundation")
         self.assertEqual(rows[1]["Price"], "19.99")
 
-        os.remove(test_file)
-        print("Test file removed✅")
+    def tearDown(self):
+        test_file = "test_user_routines.csv"
+        if os.path.exists(test_file):
+            os.remove(test_file)
+            print("Test file removed ✅")
+        else:
+            print("No test file found, no removal needed")
